backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Celery & Background Tasks
celery==5.3.6
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.0

# OAuth & Git Integration
httpx==0.26.0
authlib==1.3.0
gitpython==3.1.40

# AI & NLP
google-generativeai==0.7.2
tiktoken==0.5.2

# Code Parsing
ast-comments==1.2.1
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.3
javalang==0.13.0

# API Documentation
pyyaml==6.0.1
jsonschema==4.20.0

# Monitoring & Observability
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Email
aiosmtplib==3.0.1
emails==0.6

# Storage
boto3==1.34.34
botocore==1.34.34

# Utilities
python-slugify==8.0.1
pendulum==3.0.0
validators==0.22.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0
isort==5.13.2

# Type Stubs
types-redis==4.6.0.20240106
types-pyyaml==6.0.12.12
sqlalchemy[mypy]==2.0.25
